+++ NBT/NBTTagCompound.cs
@@ -5,7 +5,6 @@ namespace betareborn.NBT
 {
     public class NBTTagCompound : NBTBase
     {
-
         private readonly Map tagMap = new HashMap();
 
         public override void writeTagContents(DataOutput var1)
@@ -14,7 +13,7 @@ namespace betareborn.NBT
 
             while (var2.hasNext())
             {
-                NBTBase var3 = (NBTBase)var2.next();
+                NBTBase var3 = (NBTBase) var2.next();
                 writeTag(var3, var1);
             }
 
@@ -28,6 +27,7 @@ namespace betareborn.NBT
             while (true)
             {
                 NBTBase var2 = readTag(var1);
+
                 if (var2.getType() == 0)
                 {
                     return;
@@ -44,7 +44,7 @@ namespace betareborn.NBT
 
         public override byte getType()
         {
-            return (byte)10;
+            return 10;
         }
 
         public void setTag(string var1, NBTBase var2)
@@ -99,7 +99,7 @@ namespace betareborn.NBT
 
         public void setBoolean(string var1, bool var2)
         {
-            setByte(var1, (sbyte)(var2 ? 1 : 0));
+            setByte(var1, (sbyte) (var2 ? 1 : 0));
         }
 
         public bool hasKey(string var1)
@@ -109,52 +109,52 @@ namespace betareborn.NBT
 
         public sbyte getByte(string var1)
         {
-            return !tagMap.containsKey(var1) ? (sbyte)0 : ((NBTTagByte)tagMap.get(var1)).byteValue;
+            return !tagMap.containsKey(var1) ? (sbyte) 0 : ((NBTTagByte) tagMap.get(var1)).byteValue;
         }
 
         public short getShort(string var1)
         {
-            return !tagMap.containsKey(var1) ? (short)0 : ((NBTTagShort)tagMap.get(var1)).shortValue;
+            return !tagMap.containsKey(var1) ? (short) 0 : ((NBTTagShort) tagMap.get(var1)).shortValue;
         }
 
         public int getInteger(string var1)
         {
-            return !tagMap.containsKey(var1) ? 0 : ((NBTTagInt)tagMap.get(var1)).intValue;
+            return !tagMap.containsKey(var1) ? 0 : ((NBTTagInt) tagMap.get(var1)).intValue;
         }
 
         public long getLong(string var1)
         {
-            return !tagMap.containsKey(var1) ? 0L : ((NBTTagLong)tagMap.get(var1)).longValue;
+            return !tagMap.containsKey(var1) ? 0L : ((NBTTagLong) tagMap.get(var1)).longValue;
         }
 
         public float getFloat(string var1)
         {
-            return !tagMap.containsKey(var1) ? 0.0F : ((NBTTagFloat)tagMap.get(var1)).floatValue;
+            return !tagMap.containsKey(var1) ? 0.0F : ((NBTTagFloat) tagMap.get(var1)).floatValue;
         }
 
         public double getDouble(string var1)
         {
-            return !tagMap.containsKey(var1) ? 0.0D : ((NBTTagDouble)tagMap.get(var1)).doubleValue;
+            return !tagMap.containsKey(var1) ? 0.0D : ((NBTTagDouble) tagMap.get(var1)).doubleValue;
         }
 
         public string getString(string var1)
         {
-            return !tagMap.containsKey(var1) ? "" : ((NBTTagString)tagMap.get(var1)).stringValue;
+            return !tagMap.containsKey(var1) ? "" : ((NBTTagString) tagMap.get(var1)).stringValue;
         }
 
         public byte[] getByteArray(string var1)
         {
-            return !tagMap.containsKey(var1) ? [] : ((NBTTagByteArray)tagMap.get(var1)).byteArray;
+            return !tagMap.containsKey(var1) ? [] : ((NBTTagByteArray) tagMap.get(var1)).byteArray;
         }
 
         public NBTTagCompound getCompoundTag(string var1)
         {
-            return !tagMap.containsKey(var1) ? new NBTTagCompound() : (NBTTagCompound)tagMap.get(var1);
+            return !tagMap.containsKey(var1) ? new NBTTagCompound() : (NBTTagCompound) tagMap.get(var1);
         }
 
         public NBTTagList getTagList(string var1)
         {
-            return !tagMap.containsKey(var1) ? new NBTTagList() : (NBTTagList)tagMap.get(var1);
+            return !tagMap.containsKey(var1) ? new NBTTagList() : (NBTTagList) tagMap.get(var1);
         }
 
         public bool getBoolean(string var1)
@@ -167,5 +167,4 @@ namespace betareborn.NBT
             return "" + tagMap.size() + " entries";
         }
     }
-
 }
